tools/net-convert.py: support old cloudinit versions by using kwargs.
Older cloud-init versions have a bug in the signature of the
render_network_state method for netplan (bug 1685944).
The old had:
  render_network_state(target, network_state)
The fix was to change netplan's so it had the correct signature:
  render_network_state(network_state, target)

This just changes our caller to use kwargs style when invoking that
method so that it works with either the broken form or correct form.