test: Fix pyflakes complaint of unused import.
The jsonschema package is used only when available, but the lint
check thinks the import is unused across pyflakes and flake8. In
order to avoid having exceptions for both assert that the import
works right after and the import is considered used.

The '# NOQA' doesn't affect pyflakes (only flake8).

LP: #1695918