Use distro release version to determine if we use systemd in redhat spec
The typical rpm build process will examine the spec file to determine
which packages should be installed in the boot root.  This requires
the specfile to declare that it needs system.  Provide this information
by checking which version in which the rpm is being built and exporting
requirements for systemd.