net: eni route rendering missed ipv6 default route config
In some network configurations a network value of '::' and a
netmask value of '::' were used to indicate a default IPV6 gateway.
Commit d00da2d5 removed ipv6 'netmask' attributes and calculate
a prefix length value instead.  The eni route rendering failed to update
the check to use prefix value of 0 to indicate the presence of an IPV6
default route.

A broken ipv6 default route rendered like:

 post-up route add -net :: netmask :: gw 2001:4800:78ff:1b::1 || true

And with this patch, it now renders like:

 post-up route add -A inet6 default gw 2001:4800:78ff:1b::1 || true

LP: #1701097