test: fix incorrect keyid for apt repository.
The test is currently importing the incorrect keyid. It specifies
the curtin developers ppa, rather than the cloud-init ppa. On
Artful this causes failures as a check is made to verify the
correct key is imported for the ppa, whereas on previous releases
only a warning was issued.

Also, change to use a full key fingerprint.

LP: #1702717