tests: adjust locale integration test to parse default locale.
The locale integration test started failing with commit 0ef61b28.
This was just because the test's expectations on the content/formatting
of /etc/default/locale were too strict.  The change here is to read
the file as a set of shell variables and assert that the values are
set correctly.