centos: do not package systemd-fsck drop-in.
The change here fixes the build of CentOS 7 rpm.

The systemd 'drop-in' file systemd-fsck@.service.d/cloud-init.conf is
not expected to work on CentOS at the moment.  There, the
cloud-init-local.service currently runs without DefaultDependencies=no.
That would likely cause a dependency loop.

The drop-in was added for bug 1691489.  It may well be needed at a later
date in CentOS also.