resizefs: pass mount point to xfs_growfs
Supposedly it was never a feature to be able to pass a path to a block
device to xfs_growfs and have it grow the filesystem. The behavior changed
upstream recently. It is only supported to pass the mount point of a mounted
XFS filesystem. This causes breakages in cloud-init.

Upstream xfs change was commit b97815a0321072a7154ecab63e297af84066fc78.
 https://git.kernel.org/pub/scm/fs/xfs/xfsprogs-dev.git/commit/?id=b97815a0321

rhbz: rhbz: https://bugzilla.redhat.com/show_bug.cgi?id=1490505

Signed-off-by: Dusty Mabe <dusty@dustymabe.com>