CloudStack: consider dhclient lease files named with a hyphen.
A regression in 'get_latest_lease' made it ignore files starting with
'dhclient-' rather than just 'dhclient.'. The fix here is to allow those
files to be considered.

There is a lot more we could do here to better ensure that we pick the
most recent lease, but this change fixes the regression.

LP: #1717147