GCE: Fix usage of user-data.
This regressed in the rework of GCE datasource to have a main.
The fix really just stores the user-data that was read in
self.userdata_raw, rather than self.userdata.  That is consistent
with other datasources and ulitimately how it was before the refactor.

The main is updated to address the fact that user-data is binary data
and may not be able to be printed.

LP: #1717598