debian/copyright: dep5 updates, reorganize, add Apache 2.0 license.
The copyright was updated to be lintian clean and reorganized to list the
licenses at the bottom after declaring the metadata and file information.

Also cloud-init license was updated to be dual-licensed under Apache 2.0
and GPL 3.0, however, the text of the Apache 2.0 license was not added
to the debian/copyright file.

LP: #1718681