citest: show the class actual class name in results.
Tests are currently run by creating a temporary subclass of each class
and then executing it (in get_suites).  When running the tests suite
the output would contain the temporary class name.  That was less than
useful, and made batch runs almost impossible to identify which
test case had an error.

This change goes from output of:
    FAIL: test_no_warnings_in_log \
        (tests.cloud_tests.testcases.get_suite.<locals>.tmp)
To
    FAIL: test_no_warnings_in_log \
       (tests.cloud_tests.testcases.modules.ntp.TestNtp)