tests: move to using tox 1.7.5
When we moved to having tox run on tests/ and tools/ we bumped tox
to version 1.7.1.  That was an error on my part as I just checked the
version I had locally and didn't go looking for what the newest upstream
release was.

The current version as of today is 1.7.5 and 'tox -e pylint' works with
this version