tests: clean up image properties
This fixes the incorrectly named 'family' value for images as 'os'.
Families are already defined in util.py:OS_FAMILY_MAPPING and a family
is a collection of OSes.

This makes the properties function part of the super class of image as
it is only overrided by the lxd backend.