tests: Fix EC2 Platform to return console output as bytes.
The EC2 test platform uses boto, and boto decodes console output
with decode('utf-8', 'replace). It is known that Ubuntu consoles
contain non-utf8 characters, making this call lossy.

The change here is to patch the boto session to include a OutputBytes
entry in the console_output response, and then to utilize that in
console_log.

More information on problem and solution at:
   https://github.com/boto/botocore/issues/1351