net: accept network-config in netplan format for renaming interfaces
net.apply_network_config_names currently only accepts network-config
in version 1 format.  When users include a netplan format network-config the
rename code does not find any of the 'set-name' directives and does not rename
any of the interfaces.  This causes some netplan configurations to fail.

This patch adds support for parsing netplan format and extracts the needed
information (macaddress and set-name values) to allow cloud-init to issue
interface rename commands.  We know raise a RuntimeError if presented with
an unknown config format.

LP: #1709715