subp: make ProcessExecutionError have expected types in stderr, stdout.
When subp raised a ProcessExecutionError, that exception's stderr and
stdout might end up being the string '-' rather than bytes.

This mean that:
   try:
       subp(mycommand, decode=False)
   except ProcessExecutionError as e:
       pass

Would have 'e.stdout' set to '-' while the caller would expect bytes.

Also reduce the try/except block in subp to a specifically the two lines
that may raise an OSError.