tools: run-centos: git clone rather than tar.
This changes tools/run-centos to collect up your git working directory
via 'git' commands rather than just collecting the whole directory.
The reason for this is that even a clean tree that has had tox run
on it might have up to 400M of data in it.

It adds a '--dirty' flag to run-centos to collect up local changes.