ds-identify: check /writable/system-data/ for nocloud seed.
Ubuntu core seeds information to nocloud via a bind-mount of
/writable/system-data/var/lib/cloud over /var/lib/cloud.
When ds-identify runs as a systemd generator that mount is not
guaranteed to have been done.  It is guaranteed at
cloud-init-local.service time, but not generator time.

Images built with 'ubuntu-image --cloud-init=user-data-file'
would have cloud-init disabled.

The fix here is just to consider the seed dir under /writable/system-data.

LP: #1747070