subp: Fix subp usage with non-ascii characters when no system locale.
If python starts up without a locale set, then its default encoding
ends up set as ascii.  That is not easily changed with the likes of
setlocale.  In order to avoid UnicodeDecodeErrors cloud-init will
encode to bytes a python3 string or python2 basestring so that the
values passed to Popen are already bytes.

LP: #1751051