Handle global dns entries in netplan
In network config v1 format, there are dns values which are not bound to a
specific interface and do not map to the per-interface format in netplan.
To handle this case we render netplan configuration that duplicates the
DNS configuration on any interface that has a static network config.  We
avoiding interfaces which have DHCP configuration which may provide
conflicting DNS values.

LP: #1750884