DataSourceSmartOS: sdc:hostname is ignored
There are three potential sources of the hostname, one of which is
documented SmartOS's vmadm(1M) via the hostname property.  That
property's value is retrieved via the sdc:hostname key.  The other
two sources for the hostname are a hostname key in customer_metadata
and the VM's uuid (sdc:uuid).  Of these three, the sdc:hostname value
is not used in a meaningful way by DataSourceSmartOS.

This fix changes the fallback mechanism when hostname is not
specified in customer_metadata.  The order of precedence for setting
the hostname is now 1) hostname in customer_metadata,
2) sdc:hostname, then 3) sdc:uuid.

LP: #1765085