tools: Re-use the orig tarball in packages/bddeb if it is around.
If you built packages with 'bddeb', each time it would create a new
tarball with make-tarball.  If you then tried to upload two different
tarballs to launchpad (to a PPA), it would reject the second as the
orig tarball already existed.

This just supports looking in some places for a orig tarball and
re-using if it is found.