tools: Support adding a release suffix through packages/bddeb.
bddeb already supported passing in a '--release' and that would get
into the changelog line.

If you used bddeb to build packages for a PPA, and built multiple
releases, then you would get the same version for each release, and
launchpad would reject your upload.

The change here means we get a ~16.04.1 (for xenial) suffix on the
dpkg version.  If the distro-info-data package is not installed,
or the release is not known (such as the default "UNRELEASED"),
then you get no suffix.